resonate-backend/Backend-ML/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
pycryptodome
python-dotenv
httpx
pydantic